kloud: cleanup terraform data abstraction models
